+++ Sources/MapboxNavigationCore/Navigator/Internals/CoreNavigator/NavigatorStatusObserver.swift
@@ -2,9 +2,19 @@ import Foundation
 import MapboxNavigationNative
 
 class NavigatorStatusObserver: NavigatorObserver {
-    func onPrimaryRouteChanged(forPrimaryRoute primaryRoute: (any RouteInterface)?, reason: PrimaryRouteChangeReason) {}
+    func onPrimaryRouteChanged(
+        forPrimaryRoute primaryRoute: (any RouteInterface)?,
+        reason: PrimaryRouteChangeReason
+    ) {
+        // Intentionally left empty.
+    }
 
-    func onAlternativeRoutesChanged(forAlternativeRoutes alternativeRoutes: [RouteAlternative]) {}
+    func onAlternativeRoutesChanged(
+        forAlternativeRoutes alternativeRoutes: [RouteAlternative],
+        reason: AlternativeRoutesChangeReason
+    ) {
+        // Intentionally left empty. RouteAlternativesObserver is used instead.
+    }
 
     var mostRecentNavigationStatus: NavigationStatus? = nil